% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetAEE.R
\name{ce_frechet_cpp}
\alias{ce_frechet_cpp}
\title{Calculating an aysmptotical efficient estimate for the Frechet distribution}
\arguments{
\item{x}{A numeric vector.}
}
\value{
\item{theta1}{an estimated shape parameter}
\item{theta2}{an estimated scale parameter}
\item{l_n_deriv}{a vector containing the first derivative of log likelihood function}
\item{fisher_1}{a matrix of the expected fisher information divided by the number of observations}
}
\description{
Calculating an aysmptotical efficient estimate for the Frechet distribution
}
